Ability to work with tops from multiple wells
This commit adds the ability to work with tops from mulitple wells.

Previously, the app could only handle one well at a time, and tops were stored in a dictionary with format {top1: md, top2: md}.
This change stores the apps in a dataframe (or jsonified dataframe).

Tops must be loaded and stored in a tab delimited txt file that has at a minimum, the following required columns:
* UWI
* PICK
* MD

Pick data is stored in the app in the `tops-storage` div. Previously, data was input and output from this div via json.dumps or json.loads.
The new change uses df.to_json() and pd.read_json(json_data) to input and output data.

Required updates were made to the `update_pick_storage` callback to appropriately update the picks in the dataframe.
Required updates were made to the `update_figure` callback and `helper.update_picks_on_plot` to access the pick data from the dataframe to draw on the plot.
top-selector dropdown options are now determined using the unique values for picks in the dataframe.
The striplog csv callback was updated to only create the striplog for the currently active well.